+++ controllers/evaluacion.go
@@ -2,11 +2,11 @@ package controllers
 
 import (
 	"fmt"
+	"net/url"
 
 	"github.com/astaxie/beego"
-	"github.com/udistrital/planeacion_mid/services"
-	"github.com/udistrital/utils_oas/errorhandler"
-	"github.com/udistrital/utils_oas/requestresponse"
+	"github.com/udistrital/planeacion_mid/helpers"
+	evaluacionhelper "github.com/udistrital/planeacion_mid/helpers/evaluacionHelper"
 )
 
 // EvaluacionController operations for Evaluacion
@@ -16,8 +16,8 @@ type EvaluacionController struct {
 
 // URLMapping ...
 func (c *EvaluacionController) URLMapping() {
-	c.Mapping("GetPlanesPeriodo", c.GetPlanesPeriodo)
 	c.Mapping("GetEvaluacion", c.GetEvaluacion)
+	c.Mapping("GetPlanesPeriodo", c.GetPlanesPeriodo)
 	c.Mapping("PlanesAEvaluar", c.PlanesAEvaluar)
 	c.Mapping("Unidades", c.Unidades)
 	c.Mapping("Avances", c.Avances)
@@ -30,21 +30,35 @@ func (c *EvaluacionController) URLMapping() {
 // @Param	unidad 		path 	string	true		"The key for staticblock"
 // @Success 200
 // @Failure 404
-// @router /planes-periodo/:vigencia/:unidad [get]
+// @router /planes_periodo/:vigencia/:unidad [get]
 func (c *EvaluacionController) GetPlanesPeriodo() {
-	defer errorhandler.HandlePanic(&c.Controller)
+	defer func() {
+		if err := recover(); err != nil {
+			localError := err.(map[string]interface{})
+			c.Data["mesaage"] = (beego.AppConfig.String("appname") + "/" + "EvaluacionController" + "/" + (localError["funcion"]).(string))
+			c.Data["data"] = (localError["err"])
+			if status, ok := localError["status"]; ok {
+				c.Abort(status.(string))
+			} else {
+				c.Abort("404")
+			}
+		}
+	}()
 
 	vigencia := c.Ctx.Input.Param(":vigencia")
 	unidad := c.Ctx.Input.Param(":unidad")
 
-	resultado, err := services.GetPlanesPeriodo(vigencia, unidad)
+	if len(vigencia) == 0 {
+		c.Data["json"] = map[string]interface{}{"Success": false, "Status": "404", "Message": "Request containt incorrect params", "Data": nil}
+	}
+	if len(unidad) == 0 {
+		c.Data["json"] = map[string]interface{}{"Success": false, "Status": "404", "Message": "Request containt incorrect params", "Data": nil}
+	}
 
-	if err == nil {
-		c.Ctx.Output.SetStatus(200)
-		c.Data["json"] = requestresponse.APIResponseDTO(true, 200, resultado)
+	if respuesta, err := evaluacionhelper.GetPlanesPeriodo(unidad, vigencia); err == nil {
+		c.Data["json"] = map[string]interface{}{"Success": true, "Status": "200", "Message": "Successful", "Data": respuesta}
 	} else {
-		c.Ctx.Output.SetStatus(404)
-		c.Data["json"] = requestresponse.APIResponseDTO(true, 404, nil, err.Error())
+		panic(err)
 	}
 
 	c.ServeJSON()
@@ -60,42 +74,68 @@ func (c *EvaluacionController) GetPlanesPeriodo() {
 // @Failure 404
 // @router /:vigencia/:plan:/:periodo [get]
 func (c *EvaluacionController) GetEvaluacion() {
-	defer errorhandler.HandlePanic(&c.Controller)
+	defer func() {
+		if err := recover(); err != nil {
+			localError := err.(map[string]interface{})
+			c.Data["mesaage"] = (beego.AppConfig.String("appname") + "/" + "EvaluacionController" + "/" + (localError["funcion"]).(string))
+			c.Data["data"] = (localError["err"])
+			if status, ok := localError["status"]; ok {
+				c.Abort(status.(string))
+			} else {
+				c.Abort("404")
+			}
+		}
+	}()
 
+	var evaluacion []map[string]interface{}
 	vigencia := c.Ctx.Input.Param(":vigencia")
 	plan := c.Ctx.Input.Param(":plan")
 	periodoId := c.Ctx.Input.Param(":periodo")
 
-	resultado, err := services.GetEvaluacion(vigencia, plan, periodoId)
+	if len(vigencia) == 0 {
+		c.Data["json"] = map[string]interface{}{"Success": false, "Status": "404", "Message": "Request containt incorrect params", "Data": nil}
+	}
+	if len(plan) == 0 {
+		c.Data["json"] = map[string]interface{}{"Success": false, "Status": "404", "Message": "Request containt incorrect params", "Data": nil}
+	}
+	if len(periodoId) == 0 {
+		c.Data["json"] = map[string]interface{}{"Success": false, "Status": "404", "Message": "Request containt incorrect params", "Data": nil}
+	}
 
-	if err == nil {
-		c.Ctx.Output.SetStatus(200)
-		c.Data["json"] = requestresponse.APIResponseDTO(true, 200, resultado)
+	trimestres := evaluacionhelper.GetPeriodos(vigencia)
+	if len(trimestres) == 0 {
+		c.Data["json"] = map[string]interface{}{"Success": true, "Status": "200", "Message": "Successful", "Data": nil}
 	} else {
-		c.Ctx.Output.SetStatus(404)
-		c.Data["json"] = requestresponse.APIResponseDTO(true, 404, nil, err.Error())
+		i := 0
+		for index, periodo := range trimestres {
+			if periodo["_id"] == periodoId {
+				i = index
+				break
+			}
+		}
+
+		evaluacion = evaluacionhelper.GetEvaluacion(plan, trimestres, i)
+
+		c.Data["json"] = map[string]interface{}{"Success": true, "Status": "200", "Message": "Successful", "Data": evaluacion}
 	}
 
 	c.ServeJSON()
 }
 
 // Get Planes A Evaluar ...
-// @Title GetPlanesAEvaluar
+// @Title PlanesAEvaluar
 // @Description get Planes que se pueden evaluar
 // @Success 200
 // @Failure 404
 // @router /planes/ [get]
 func (c *EvaluacionController) PlanesAEvaluar() {
-	defer errorhandler.HandlePanic(&c.Controller)
-	fmt.Println("Entered HomeHandler")
-	resultado, err := services.PlanesAEvaluar()
+	fmt.Println("entrando a planes a evaluar")
+	defer helpers.ErrorController(c.Controller, "EvaluacionController")
 
-	if err == nil {
-		c.Ctx.Output.SetStatus(200)
-		c.Data["json"] = requestresponse.APIResponseDTO(true, 200, resultado)
+	if datos, err := evaluacionhelper.PlanesAEvaluar(); err == nil {
+		c.Data["json"] = map[string]interface{}{"Success": true, "Status": "200", "Message": "Successful", "Data": datos}
 	} else {
-		c.Ctx.Output.SetStatus(404)
-		c.Data["json"] = requestresponse.APIResponseDTO(true, 404, nil, err.Error())
+		panic(map[string]interface{}{"funcion": "PlanesAEvaluar", "err": err, "status": "404", "message": "Error obteniendo los planes a evaluar"})
 	}
 	c.ServeJSON()
 }
@@ -109,20 +149,21 @@ func (c *EvaluacionController) PlanesAEvaluar() {
 // @Failure 404
 // @router /unidades/:plan:/:vigencia [get]
 func (c *EvaluacionController) Unidades() {
-	defer errorhandler.HandlePanic(&c.Controller)
+	defer helpers.ErrorController(c.Controller, "EvaluacionController")
 
 	plan := c.Ctx.Input.Param(":plan")
 	vigencia := c.Ctx.Input.Param(":vigencia")
 
-	resultado, err := services.Unidades(plan, vigencia)
-
-	if err == nil {
-		c.Ctx.Output.SetStatus(200)
-		c.Data["json"] = requestresponse.APIResponseDTO(true, 200, resultado)
+	if nombrePlan, err := url.QueryUnescape(plan); err == nil {
+		if data, err := evaluacionhelper.GetUnidadesPorPlanYVigencia(nombrePlan, vigencia); err == nil {
+			c.Data["json"] = map[string]interface{}{"Success": true, "Status": "200", "Message": "Successful", "Data": data}
+		} else {
+			panic(map[string]interface{}{"funcion": "Unidades", "err": err, "status": "404", "message": "Error obteniendo las unidades del plan y la vigencia dados"})
+		}
 	} else {
-		c.Ctx.Output.SetStatus(404)
-		c.Data["json"] = requestresponse.APIResponseDTO(true, 404, nil, err.Error())
+		panic(map[string]interface{}{"funcion": "Unidades", "err": err, "status": "404", "message": "Error obteniendo las unidades del plan y la vigencia dados"})
 	}
+
 	c.ServeJSON()
 }
 
@@ -136,20 +177,20 @@ func (c *EvaluacionController) Unidades() {
 // @Failure 404
 // @router /avance/:plan:/:vigencia/:unidad [get]
 func (c *EvaluacionController) Avances() {
-	defer errorhandler.HandlePanic(&c.Controller)
+	defer helpers.ErrorController(c.Controller, "EvaluacionController")
 
 	plan := c.Ctx.Input.Param(":plan")
 	vigencia := c.Ctx.Input.Param(":vigencia")
 	unidad := c.Ctx.Input.Param(":unidad")
 
-	resultado, err := services.Avances(plan, vigencia, unidad)
-
-	if err == nil {
-		c.Ctx.Output.SetStatus(200)
-		c.Data["json"] = requestresponse.APIResponseDTO(true, 200, resultado)
+	if nombrePlan, err1 := url.QueryUnescape(plan); err1 == nil {
+		if data, err2 := evaluacionhelper.GetAvances(nombrePlan, vigencia, unidad); err2 == nil {
+			c.Data["json"] = map[string]interface{}{"Success": true, "Status": "200", "Message": "Successful", "Data": data}
+		} else {
+			panic(map[string]interface{}{"funcion": "Avances", "err": err2, "status": "404", "message": "Error obteniendo los avances"})
+		}
 	} else {
-		c.Ctx.Output.SetStatus(404)
-		c.Data["json"] = requestresponse.APIResponseDTO(true, 404, nil, err.Error())
+		panic(map[string]interface{}{"funcion": "Avances", "err": err1, "status": "404", "message": "No se pudo obtener el nombre"})
 	}
 
 	c.ServeJSON()
