+++ routers/commentsRouter_controllers.go
@@ -72,8 +72,8 @@ func init() {
 
     beego.GlobalControllerRouter["github.com/udistrital/planeacion_mid/controllers:EvaluacionController"] = append(beego.GlobalControllerRouter["github.com/udistrital/planeacion_mid/controllers:EvaluacionController"],
         beego.ControllerComments{
-            Method: "GetPlanesPeriodo",
-            Router: "/planes-periodo/:vigencia/:unidad",
+            Method: "PlanesAEvaluar",
+            Router: "/planes/",
             AllowHTTPMethods: []string{"get"},
             MethodParams: param.Make(),
             Filters: nil,
@@ -81,8 +81,8 @@ func init() {
 
     beego.GlobalControllerRouter["github.com/udistrital/planeacion_mid/controllers:EvaluacionController"] = append(beego.GlobalControllerRouter["github.com/udistrital/planeacion_mid/controllers:EvaluacionController"],
         beego.ControllerComments{
-            Method: "PlanesAEvaluar",
-            Router: "/planes/",
+            Method: "GetPlanesPeriodo",
+            Router: "/planes_periodo/:vigencia/:unidad",
             AllowHTTPMethods: []string{"get"},
             MethodParams: param.Make(),
             Filters: nil,
